clean up stories
