+++ src/js/components/List/stories/key render.js
@@ -4,7 +4,7 @@ import { storiesOf } from '@storybook/react';
 import { Grommet, Box, List, Text } from 'grommet';
 import { grommet } from 'grommet/themes';
 
-export const locations = [
+const locations = [
   'Boise',
   'Fort Collins',
   'Los Gatos',
@@ -12,15 +12,11 @@ export const locations = [
   'San Francisco',
 ];
 
-export const data = [];
-
+const data = [];
 for (let i = 0; i < 40; i += 1) {
   data.push({
     entry: `entry-${i + 1}`,
     location: locations[i % locations.length],
-    date: `2018-07-${(i % 30) + 1}`,
-    percent: (i % 11) * 10,
-    paid: ((i + 1) * 17) % 1000,
   });
 }
