+++ src/js/components/List/stories/secondaryKey.js
@@ -18,9 +18,6 @@ for (let i = 0; i < 40; i += 1) {
   data.push({
     entry: `entry-${i + 1}`,
     location: locations[i % locations.length],
-    date: `2018-07-${(i % 30) + 1}`,
-    percent: (i % 11) * 10,
-    paid: ((i + 1) * 17) % 1000,
   });
 }
