+++ src/js/components/List/stories/selection.js
@@ -4,23 +4,11 @@ import { storiesOf } from '@storybook/react';
 import { Grommet, Box, List } from 'grommet';
 import { grommet } from 'grommet/themes';
 
-export const locations = [
-  'Boise',
-  'Fort Collins',
-  'Los Gatos',
-  'Palo Alto',
-  'San Francisco',
-];
-
-export const data = [];
+const data = [];
 
 for (let i = 0; i < 40; i += 1) {
   data.push({
     entry: `entry-${i + 1}`,
-    location: locations[i % locations.length],
-    date: `2018-07-${(i % 30) + 1}`,
-    percent: (i % 11) * 10,
-    paid: ((i + 1) * 17) % 1000,
   });
 }
