+++ src/js/components/List/stories/theme.js
@@ -5,7 +5,13 @@ import { Grommet, Box, List } from 'grommet';
 import { grommet } from 'grommet/themes';
 import { deepMerge } from 'grommet/utils';
 
-import { locations } from './data';
+export const locations = [
+  'Boise',
+  'Fort Collins',
+  'Los Gatos',
+  'Palo Alto',
+  'San Francisco',
+];
 
 const theme = deepMerge(grommet, {
   list: {
